Refactored folder structure for easy maintenance
