+++ Documentation/userGuide/source/download.rst
@@ -11,6 +11,17 @@ See the release notes for additional information on the FMI version 2.0 for Co-S
 
 To install EnergyPlusToFMU, follow the section :doc:`installation`.
 
+Release 3.0.1 (xxx xxx, 2020)
+---------------------------------
+
+Download `EnergyPlusToFMU-v3.0.1.zip <https://github.com/lbl-srg/EnergyplusToFMU/releases/download/v3.0.1/EnergyPlusToFMU-v3.0.1.zip>`_.
+
+**Release notes**
+
+This version fixes a bug which was causing an incorrect parsing of the Runperiod.
+This addresses `issue#30 <https://github.com/lbl-srg/EnergyPlusToFMU/issues/30/>`_ on github.
+
+
 Release 3.0.0 (April 17, 2020)
 ---------------------------------
