+++ SourceCode/fmu-export-prep-shared/fmu-export-idf-data.h
@@ -16,9 +16,9 @@
 //#include <fstream>
 //#include <string>
 
-#include "../../read-ep-file/ep-idd-map.h"
-#include "../../read-ep-file/fileReaderData.h"
-#include "../../read-ep-file/fileReaderDictionary.h"
+#include "../read-ep-file/ep-idd-map.h"
+#include "../read-ep-file/fileReaderData.h"
+#include "../read-ep-file/fileReaderDictionary.h"
 
 
 //-- Read and store data from EnergyPlus IDF file.
