+++ SourceCode/fmu-export-prep-shared/fmu-export-write-vars-cfg.cpp
@@ -15,8 +15,8 @@ using std::endl;
 
 #include "fmu-export-write-vars-cfg.h"
 
-#include "../../utility/file-help.h"
-#include "../../utility/xml-output-help.h"
+#include "../utility/file-help.h"
+#include "../utility/xml-output-help.h"
 
 
 //--- File-scope constants.
